% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psp_ac_codes}
\alias{psp_ac_codes}
\title{Agricultural Accounts items and categories}
\format{
a dataframe with 3 columns and 100 rows:
\describe{
\item{ac_id}{an ID from the accounts Final Value Table}
\item{ac_item}{an item description for an ac_id}
\item{psp_cat}{an aggregate product category}
}
}
\usage{
psp_ac_codes
}
\description{
Agricultural Accounts items and categories
}
\keyword{datasets}
