% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{read_accounts}
\alias{read_accounts}
\title{Read an agricultural accounts table}
\usage{
read_accounts(cells)
}
\arguments{
\item{cells}{a tidyxl::xlsx_cells output}
}
\value{
a dataframe in tidy format
}
\description{
Extract data from a tidyxl::xlsx_cells output
}
\examples{
\dontrun{
ac <- system.file("extdata/accounts.xlsx", package = "psp") |>
tidyxl::xlsx_cells()

read_accounts(ac)
}
}
