% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{url_sut}
\alias{url_sut}
\title{URL for Supply and Use tables}
\usage{
url_sut()
}
\value{
The location of ONS Supply and Use tables
}
\description{
URL for Supply and Use tables
}
\examples{
\dontrun{
url <- url_sut()
}
}
