\name{psp-package}
\alias{psp}
\docType{package}
\title{

    Parameter Space Partitioning MCMC for Global Model Evaluation

}

\description{

Implements global model evaluations for formal computational models in the
cognitive sciences.

}

\keyword{computational modelling; parameter space partitioning; model evaluation}

\author{

    Lenard Dome

    Maintainer: Lenard Dome \email{lenarddome@protonmail.ch}

}
