\name{psp-package}
\alias{psp-package}
\docType{package}
\title{

    Parameter Space Partitioning MCMC for Global Model Evaluation

}

\description{

Implements an n-dimensional parameter space partitioning algorithm for evaluating
the global behaviour of formal computational models as described by Pitt, Kim,
Navarro and Myung (2006) <doi:10.1037/0033-295X.113.1.57>.

Please cite the package in publications. Use \code{citation("psp")}.

}

\keyword{computational modelling; parameter space partitioning; model evaluation}

\author{

    Lenard Dome

    Maintainer: Lenard Dome \email{lenarddome@gmail.com}

}

\references{
    Pitt, M. A., Kim, W., Navarro, D. J., & Myung, J. I. (2006). Global model analysis by parameter space partitioning. Psychological Review, 113(1), 57.
}
